+++ Scripts/Spells/Autres-Inutilises/Divins-VieuxTemrael/Shamanisme/Fetichisme.cs
@@ -122,7 +122,7 @@ namespace Server.Mobiles
             if (Target.Weapon != null)
             {
                 BaseWeapon weapon = (BaseWeapon)Target.Weapon;
-                SetDamage(weapon.MinDamage - 5, weapon.MaxDamage + 5);
+                SetDamage((int)weapon.MinDamage - 5, (int)weapon.MaxDamage + 5);
             }
             else
             {
